requirements.txt
python-dotenv==1.0.0
flask==2.3.2
Flask-WTF==1.1.1
WTForms==3.0.1
openai==0.27.8
